Fixing (missed) date and time to_f conversions (related to neildecapia/xapit@331e3d1)

+++ lib/xapit/indexers/abstract_indexer.rb
@@ -47,9 +47,9 @@ module Xapit
       @blueprint.field_attributes.map do |name|
         [member.send(name)].flatten.map do |value|
           if value.kind_of? Time
-            value = value.to_i
+            value = value.to_f
           elsif value.kind_of? Date
-            value = value.to_time.to_i
+            value = value.to_time.to_f
           end
           "X#{name}-#{value.to_s.downcase}"
         end
